intial commmit: Script to sample google trends data
Google presenets a random selection of google trends data (older than 36 hours). For keywords with low search volume it seems like the sampling is to small to accuractly get the search history trend. In order to get many different random samples a sliding window is used pulling the same data for many different time periods, averaging and normalizing the results